requirements.txt
uiautomator2>=2.16.0
Pillow>=9.0.0
requests>=2.28.0
python-dotenv>=0.19.0
openai>=1.0.0
pydantic>=2.0.0

